@@ -13,17 +13,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^20.1.6",
+    "@angular/animations": "^20.3.2",
     "@angular/cdk": "^19.2.19",
-    "@angular/common": "^20.1.6",
-    "@angular/compiler": "^20.1.6",
-    "@angular/core": "^20.1.6",
+    "@angular/common": "^20.3.2",
+    "@angular/compiler": "^20.3.2",
+    "@angular/core": "^20.3.2",
     "@angular/fire": "^20.0.1",
-    "@angular/forms": "^20.1.6",
+    "@angular/forms": "^20.3.2",
     "@angular/material": "^19.2.19",
-    "@angular/platform-browser": "^20.1.6",
-    "@angular/platform-browser-dynamic": "^20.1.6",
-    "@angular/router": "^20.1.6",
+    "@angular/platform-browser": "^20.3.2",
+    "@angular/platform-browser-dynamic": "^20.3.2",
+    "@angular/router": "^20.3.2",
     "@capacitor-community/firebase-analytics": "^7.0.1",
     "@capacitor/android": "^7.4.3",
     "@capacitor/cli": "^7.4.3",
@@ -33,7 +33,7 @@
     "@ionic/angular": "^8.0.0",
     "bootstrap": "^5.3.6",
     "fire": "^0.55.1",
-    "firebase": "^12.1.0",
+    "firebase": "^12.3.0",
     "ionicons": "^7.0.0",
     "rxjs": "~7.8.0",
     "swiper": "^11.2.8",
@@ -47,10 +47,10 @@
     "@angular-eslint/eslint-plugin-template": "^19.0.0",
     "@angular-eslint/schematics": "^19.0.0",
     "@angular-eslint/template-parser": "^19.0.0",
-    "@angular/build": "^20.1.5",
-    "@angular/cli": "^20.1.5",
-    "@angular/compiler-cli": "^20.1.6",
-    "@angular/language-service": "^20.1.6",
+    "@angular/build": "^20.3.3",
+    "@angular/cli": "^20.3.3",
+    "@angular/compiler-cli": "^20.3.2",
+    "@angular/language-service": "^20.3.2",
     "@ionic/angular-toolkit": "^12.0.0",
     "@types/google.maps": "^3.58.1",
     "@types/jasmine": "~5.1.0",
